tests/wpt/web-platform-tests/css/css-transforms/css-transform-property-existence.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>CSS Transforms Test: property existence</title>
    <link rel="author" title="Intel" href="http://www.intel.com">
    <link rel="help" href="http://www.w3.org/TR/css-transforms-1/#propdef-transform">
    <meta name="flags" content="dom">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="test"></div>
    <div id="log"></div>
    <script>
      test(function() {
        assert_not_equals(document.getElementById("test").style.transform, undefined, "expect transform is not undefined");
      }, "Check the existence of transform.");

      test(function() {
        assert_own_property(document.getElementById("test").style, "transform-origin", "expect transform-origin");
      }, "Check the existence of transform-origin.");

      test(function() {
        assert_own_property(document.getElementById("test").style, "transform-style", "expect transform-style");
      }, "Check the existence of transform-style.");

      test(function() {
        assert_own_property(document.getElementById("test").style, "perspective", "expect perspective");
      }, "Check the existence of perspective.");

      test(function() {
        assert_own_property(document.getElementById("test").style, "perspective-origin", "expect perspective-origin");
      }, "Check the existence of perspective-origin.");

      test(function() {
        assert_own_property(document.getElementById("test").style, "backface-visibility", "expect backface-visibility");
      }, "Check the existence of backface-visibility.");
    </script>
  </body>
</html>
